:truck: move api  tests to v1 folder
