+++ src/objects/tests/v1/test_geo_headers.py
@@ -1,5 +1,3 @@
-from django.urls import reverse
-
 from rest_framework import status
 from rest_framework.test import APITestCase
 
@@ -8,7 +6,8 @@ from objects.token.constants import PermissionModes
 from objects.token.tests.factories import PermissionFactory
 from objects.utils.test import TokenAuthMixin
 
-from .constants import GEO_READ_KWARGS, POLYGON_AMSTERDAM_CENTRUM
+from ..constants import GEO_READ_KWARGS, POLYGON_AMSTERDAM_CENTRUM
+from .utils import reverse
 
 OBJECT_TYPES_API = "https://example.com/objecttypes/v1/"
