+++ src/objects/tests/v1/test_object_api.py
@@ -1,8 +1,6 @@
 import json
 from datetime import date, timedelta
 
-from django.urls import reverse
-
 import requests_mock
 from freezegun import freeze_time
 from rest_framework import status
@@ -18,8 +16,9 @@ from objects.token.constants import PermissionModes
 from objects.token.tests.factories import PermissionFactory
 from objects.utils.test import TokenAuthMixin
 
-from .constants import GEO_WRITE_KWARGS
-from .utils import mock_objecttype_version, mock_service_oas_get
+from ..constants import GEO_WRITE_KWARGS
+from ..utils import mock_objecttype_version, mock_service_oas_get
+from .utils import reverse
 
 OBJECT_TYPES_API = "https://example.com/objecttypes/v1/"
