+++ src/objects/tests/v1/test_object_api_fields.py
@@ -1,7 +1,5 @@
 from datetime import date
 
-from django.urls import reverse
-
 from rest_framework import status
 from rest_framework.test import APITestCase
 
@@ -13,6 +11,8 @@ from objects.core.tests.factores import (
 from objects.token.constants import PermissionModes
 from objects.token.tests.factories import PermissionFactory
 from objects.utils.test import TokenAuthMixin
+
+from .utils import reverse
 
 OBJECT_TYPES_API = "https://example.com/objecttypes/v1/"
